feat:信用卡申请接口，激活接口，获取信用卡信息

+++ src/main/java/com/card/mapper/UserMapper.java
@@ -13,7 +13,7 @@ import java.util.Map;
 @Component
 public interface UserMapper {
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(long id);
 
     long insert(User record);
